Implementation of Pipeline object (#95)
Pipeline is initialized by moving in an object on which it creates a ready future and the only method it exposes is the call operator Pipeline<T>::operator()() that returns an hpx::future.

Internally, at every call, it chains all the futures, so that they will become ready in a serialized and ordered way. In this chain, the object given at initialization time is moved (i.e. std::move) from future to future as soon as they get ready, by means of the Wrapper object.
The Wrapper object is used to contain the real object (the moved one) and to make ready the next future as soon as it is destroyed. A reference to the object can be accessed through the call operatorWrapper<T>::operator()().