API and directory structure of algorithms: Example with Cholesky (#146)
Introduced new API and directory structure for Cholesky.

The new API Introduce categories for the operations:
- factorization
- solver
- inverse
- multiplication
- ...

The new API allow the choice of the Backend (currently only MC is available but other will be added). Moreover allow a Default Backend can be implemented which choose the best backend available for a specific problem.