Add column/row-major iterators and iterables (#199)
`IteratorRange2D` and `IterableRange2D` interoperate with `Index2D` and `Size2D` and provide a way to iterate a 2D range in column-major order with a range-based for loop. The goal is to minimize chances for errors by reducing boilerplate and avoiding intermediary iteration loop variables. 

A number of function wrappers `iterateRange2D()` are provided to deduce types in constructor calls to `IterableRange2D`.