Make sure no yielding happens while handling exceptions (#269)
See STEllAR-GROUP/hpx#4894 for more details. The short version is that starting a catch block on one worker thread and resuming on another is something that does not work at least with GCC/clang and libstdc++/libc++ (this seems to be handled correctly on Windows). promise::set_exception requires a lock which may yield.