Algorithm gen2std eigenproblem (lower case) (#261)
Implementation of the algorithm for the transformation from generalized to standard eigenproblem. The algorithm is implemented for CPU, only on local memory so far. This method solves the equation: B= L^(-1) A L^(-H), where L is a lower triangular matrix (obtained with a Cholesky decomposition and A is a Hermitian square matrix).

Tests, using analytical expressions, are implemented both for this algorithm and for the tile wrapper of the HEGST LAPACK function (for this specific case).

Also, HEMM and HERK BLAS wrapper functions have been added along with their tests.

Two functions for printing elements of matrices or tiles have been implemented.
The documentation has been fixed in a few files.