Matrix helper for tests (#291)
The aim of this PR is introducing a simple data-structure for a Matrix, useful during tests writing.

For this reason I would go for keeping it in the dlaf::test namespace, not making it part of our library.

The main idea is to have a MatrixLocal that has these main properties:
- it is not async, so it does not have to deal with futures;
- it is not thread-safe;
- it is fully locally stored;
- it should be possible to work with it in terms of tiles or directly elements: this double-abstraction over it allows to keep a direct connection with a distributed matrix (if it has been "gathered" from it) but at the same time it is easy to access (address) of elements for inter-operability with BLAS/LAPACK;

At the moment I wouldn't mind about having it device portable (e.g. stored on GPU or CPU). The philosophy can be, as soon as we need it, we can improve it.

I already have some helper functions that I'm going to add soon. Some examples:

- set, similar to what it is already available for dlaf::matrix::Matrix
- all_gather that allows to collect locally a distributed dlaf::matrix::Matrix