Library initialization/finalization (#312)
Note that this builds on #234, and thus the diff looks a bit larger than it really is. I've squashed the current changes into the last commit.

This addresses #100/#306.

This is a draft for a couple things:

    dlaf::initialize/finalize for (at this point) initializing static CUDA stream/cuBLAS handle pools, but this could be useful for other things as well.
    A way to generically use functionality for different backends (with different executors) (not really new, but just applied to a new thing: copy).

As a bonus, it also demonstrates the way I've used so far for extending tile lifetimes generically. I can eventually split this out into a separate PR, but it's related as it makes use of the possibility to get executors for different backends.

To expand on what this includes right now:

    dlaf::initialize currently takes argc/argv but does nothing with them. It initializes the static pools for the GPU backend. I could see DLAF eventually getting some command line options for controlling e.g. the number of streams per worker thread. Do we want to do that now? If yes, with hpx::program_options? With something else?
    Currently the only thing I've added is getCopyExecutor which returns different types of executors for CPU and GPU (note that the only reason why it's specifically copy, not e.g. normal priority, is because I don't have the cuBLAS executor forward to the superclass async_execute at the moment. I'm looking into making this work.) For this to actually be generic I've added some helper functionality.

The first helper is a macro for creating callable objects for a particular function name. The callable object simply passes through any and all arguments it gets and dispatches to the actual function. This doesn't have to be done with a macro, but I think in many cases it simplifies having to create the wrapper manually.

The second helper is unwrapExtendTiles (name needs some bikeshedding). This is essentially a custom version of hpx::util::unwrapping which does the unwrapping, but also returns a tuple of the arguments to extend the lifetimes of the tiles. Currently it simply returns a tuple of all the arguments, but this could perhaps be made to return only the tile arguments. I'm not sure yet how to do that.