computeTFactor: API change (#331)
this PR proposes two major API changes:

- taus are passed as shared_future<vector<T>> and not as vector<shared_future<T>>;
- pass a future<Tile> instead of a Matrix as storage for the output

The former suggests better the idea that all the taus for this panel are considered together instead as a results of single columns. This reduces the atomicity of the operation, but the granularity at column level is not interesting because it leads to numerous "microtasks" (not advisable due to overhead and management by the scheduling).

The latter allows to use any Tile (of the right size) from whatever Matrix and in the preferred position (not strictly the top-left one). Since there is no more the Matrix that manages the "pipelining" of the accesses to the tile, it required to do it inside the algorithm (keeping/returning the future from tasks).