Add default Device and Backend enumerations (#338)
Simply adds Device::Default and Backend::Default to the Device and Backend enums, and they are equal to MC/CPU when CUDA is disabled and GPU/GPU when CUDA is enabled. This helps avoiding ifdefs in user code (i.e. miniapps) to choose the most "capable" backend/device that is available.