New MPI API (#344)
This is the new MPI API based on non-blocking MPI communication.
Different approaches were considered as follows:
1) yielding
2) polling

which work with non-blocking MPI communication. The difference between the two approaches is in the internal mechanism for testing the MPI_Request for completion. 1. tests each request individually (with MPI_Test()), 2. places the request in a global array which is then tested in bulk (with MPI_Testall()) whenever a task yields or completes.

The MPI executor implemented in this PR supports both the async() and dataflow() HPX APIs. The APIs is such that the last parameter of any function scheduled using the executor is required to be MPI_Request* req_ptr (see kernels.h).

common::PromiseGuard<Communicator> pcomm is necessary to supply the communicator AND to impose order between tasks launched on that communicator. The purpose of this object hasn't changed.

Attempts to manage MPI task ordering within the MPI executor proved to be challenging and fragile due to internal details of the dataflow API. The dataflow_finalize() function implemented in the executor is often executed in a continuation which requires that the executor is copied and complicates the lifetime-management of the ordering mechanism within the executor considerably.

The switch between approaches 1. and 2. was initially implemented at compile-time. A few days ago I changed the switch to a runtime parameter. Measurements showed no performance difference either way. One reason for the switch was to reduce the number of explicit template instantiations that were needed for each algorithm. Another is the avoid having to introduce additional struct within api.h to manage an extra template parameter.