Select a range of tiles (for creating a barrier) (#335)
This is a proposal for adding a way to get a set of tiles from a Matrix, in readwrite or readonly mode.

In particular, two functions are proposed, select and select_read, which return a vector of futures or shared_futures of Tile respectively.

With the return vector, it would be possible to do something like

Matrix m;

// create a barrier
hpx::when_all(select_read(m, iterate_range2d(m.nrTiles())).get();

// select a range of tiles, e.g. a panel, and then operate on them
hpx::dataflow(do_something_on_panel, select(m, iterate_range2d({1, 1}, {2, 3}));

These extension to the API can play an important role in algorithms in order to avoid problems like micro-tasking.

Just for giving a bit of context about latter case with an example, in redutionToBand algorithm it is necessary to work with a single matrix column spread across multiple tiles (and the next column has to wait for the first column result). The micro-tasking problem arises due to code like this

for (auto index : panel_range) {
 ...
 hpx::dataflow(work_on_single_column, matrix(index));
}

while we can "reverse-out" that part of the algorithm with

hpx::dataflow([](std::vector<hpx::future<Tile>> tiles){
  for (auto& tile : tile)
    work_on_single_column(tile);
}, select(matrix, panel_range));

Getting a single HPX task that internally has a for-loop that groups a set of small operations, instead of having a moltitude of micro-tasks which may overload the scheduler and on which the overhead may easily become unsustainable.

Clearly, this is not possible everywhere because it creates a barrier and it "locks/unlocks" dependencies all at once, but if there is anyway an implicit dependency, it helps improving the performances.