Mirror matrix (#321)
For #319 there is a need for transferring matrices between host and device. Usually this needs to not only be done once, but the copies need to be synchronized. This issue is meant to discuss options.

What I would propose is a MatrixMirror type that handles the copying. An important requirement is that the mirror does not make any copies if the source and target devices are the same. In that case the mirror would act exactly as a reference to the original matrix. Sketch:

template <class T, Device Target, Device Source>
class MatrixMirror {
    // Store the source matrix as a reference. The source can still be accessed.
    // This creates a copy (if needed) on the target device.
    MatrixMirror(Matrix<T, Source> const&);
    // This copies matrix back from the target to the source.
    ~MatrixMirror();

    // Synchronize explicitly from target to source.
    void syncTargetToSource();
    // Synchronize explicitly from source to target.
    void syncSourceToTarget();

    // Access the target (copy) matrix.
    future<TileType> operator()(LocalTileIndex);
    shared_future<TileType> read(LocalTileIndex);
    // And the same for global tile indices...
};

In the last commit on this branch I've only implemented a HostMatrix (should really have mirror in the name...) which is a mirror of a source matrix on Device::CPU memory. I think the more general formulation from above is probably a better choice. Typedefs could be introduced for the common case of the source or target being Device::CPU. This would need to conform to the MatrixLike template template parameters that are used in a few places.

Note that an independent feature would be to implement set for Device::GPU matrices as well. I'm doing this first for a few reasons: 1. I'm not convinced about the CUDA executor in HPX being the right choice, and 2. the MatrixMirror covers that case as well (although with copies), and 3. gtest checks should be done on the host, so I would anyway need this feature to transfer device matrices to the host. This feature allows doing that generically.

Unfortunately builds on top of #319 because that's where I get to test this, but the feature itself is in principle independent of it.