Panel and common communication patterns (#309)
This PR tries to address a common need in writing algorithms about workspaces.
In writing algorithms, it is common using additional panel workspaces for storing partial results or operands that are shared among multiple ranks for usage in local computations. The workspace represents a set of tiles locally stored, so it may be thought as a Matrix, non-distributed, with few additional features.

The main feature is that each tile in the workspace can be either internally allocated or mapped to an external one (e.g. from another Matrix). This is useful both for using local available tiles without copying their content and for coordinate mapping, (e.g. making a tile reachable at another coordinate, such as in panel transposition).