bug fix: wrong size of T
due to the reduced number of reflectors, T were ending up with the
wrong size with respect to the full size that it is used for the
second phase of the algorithm working on the update of the trailing
matrix. This will be changed in the future, because it is not needed
to work on the full tile at that point, but for the moment it is easier
like this and it should not impact dramatically performances.