Improve Panel API and functionality with global range (#384)
Previously the range for the panel had to be specified through local indices, often leaving the burden to the user to convert from the global index to the local one.

Now, the range is specified with global coordinates. In addition to improving its usage from the user point of view, it also opens to a better versatility. For instance, by having access to the global range coordinates, it is possible to retrieve the position of the panel in the global coordinate system of the matrix, which was not possible with just local range coordinates.