Tile utility to access subtiles. (#389)
splitTile returns a tile or a vector of tiles which represent the subtiles:

in the case of const tiles all the tiles (including the original full tile) are ready at the same time. The next tile extracted from the matrix will be ready only when all the subtiles and the full tile go out of scope.

in the case of modifiable tiles the subtiles has to be disjoint (to avoid race conditions), therefore all the subtiles will become ready at the same time. Moreover the future passed to splitTile is replaced by a new future containing the full tile that will become ready when all the subtiles will go out of scope. Finally the next tile extracted from the matrix will be ready only when the full tile goes out of scope.