algo: triangular solver LLT (distributed) (#409)
Implementation of the distributed Triangular Solver Left Lower Transposed.
Note:
keep_future is problematic with transform<GPU> because it releases
the future during the submission of the work but then it does not
keep it alive till the end of the async operation. Not using it
results in getting the contained value, which instead will be kept
alive (it does not get unwrapped as the future).