Pipeline fixes: non-blocking d'tor (#423)
bout the deadlock and the temporary workaround with hpx::threads::get_thread_manager().wait().

When running tests on PizDaint GPU partition, which run on a single node, with 6 ranks we get 12 cores / 6 ranks = 2 cores/rank. Considering that 1 thread is dedicated to the MPI pool for asynchronous communication (mpi pool), in this configuration we end up having a single core for the default pool.

On the default pool runs both the computations tasks of the algorithms, and also the so called hpx_main, the one which we generally use just for scheduling things.

Let's see what was creating the deadlock in the tests.
The algorithm A0 exits, since the call is not blocking anymore, but after it the test waits for the local tiles for checking the results. Anyway, this does not imply that all algorithm tasks finished, for instance there may be some computation (e.g. on a panel) that are not relevant for the local tiles but does for the other ranks, i.e. they have to be computed and then communicated.
So, we have some algorithm tasks not yet completed/scheduled, but the local check can complete and go to the next test config starting the algorithm A1, which as first thing does the MPI_Comm_dup in the hpx_main task.

And here it is the problem: MPI_Comm_dup is a collective AND a blocking call, which means that does not complete until all ranks do the same call. But, we are in a condition where:

    ranks which already completed A0 are stuck in MPI_Comm_dup and cannot do anything else until all other ranks reach the same point; they are stuck because the only core available in default is blocked. This implies that any other tasks for A0 not yet complete cannot be run, and in turn they cannot unlock the related communications that would permit the advancement of other ranks;
    ranks which are still executing A0 are waiting data, that will be never communicated, and so they will never reach the MPI_Comm_dup.

And here it is the deadlock. This is what we found out with gen_to_std, but a similar scenario happens for reduction_to_band, where instead of having the MPI_Comm_dup we have an MPI collective call (i.e. MPI_AllGather) for the check phase of the test, that creates exactly the same problem.

The workaround at the moment is to wait for all the tasks before scheduling a collective blocking MPI call. When the algorithm exists, it already scheduled in HPX all the tasks, so we are making the algorithm blocking again.

Note: the barrier we add with hpx::threads::get_thread_manager().wait() is stronger than the Pipeline blocking d'tor. Indeed, the Pipeline blocking d'tor was waiting for the last MPI task to be posted in MPI, not even that it was finished. Moreover, with this kind of barrier now we are waiting for anything to complete before continuing (i.e. we wait also for computations, not just mpi).