Remove pointer to promise by variant in Tile. (#439)
As the tile promise is optional within a Tile the member was declared as a unique ptr (std::optional was not available in C++14). Subtiles added two extra shared futures and therefore std::variant became the better option as only one of object can be present at a given time.

However an object cannot contain a optional/variant of an object of its type, therefore a small refactoring was needed.