Panel: add setRangeStart for non-tile aligned offset (#478)
In particular, for what concerns the Panel, the sub-tile requires that:
- first global tile of the panel may be incomplete due to a sub-tile offset;
- panel size along axis dimension has to be "constrained" to sub-tile size;

This PR targets just the first point, since for the second one the choice has been to exploit the non-strictness of the API.

We used to give the actual Distribution of the matrix which the panel supports to construct it. Now, in order to work with sub-tiles, we will pass a Distribution object created ad-hoc, with a blocksize customized to represent the sub-tile size.