algo: backtransform tridiagonal to band, adaptation to sub-band (#481)
This PR introduces the possibility to compute the backtransformation from tridiagonal to a band matrix, where the band size is a divisor of the blocksize.

This removes the constraint about having the band size equal to the blocksize, which will be mainly useful with the next GPU version of the algorithm. Indeed, with GPU block sizes will be bigger, but with this it is possible to keep the band size smaller, reducing the overload on the entire eigensolver pipeline.