Add option to use precompiled headers (#495)
Fixes #452.

While testing on daint this removes up to 5 seconds per translation unit. The initial compilation of the precompiled headers takes ~15 seconds.

Currently the default for the option is off, and I think at least CI should not have this enabled since this can hide the lack of inclusion of headers in the precompiled headers. My suggestion is that we keep it off by default for now, and you all test it in your own setups and if we later conclude that everyone can use this without problems we can enable it by default.