Use pika's transform_mpi, polling and stream throttling support
completion_modes:
pika supports different completion modes that may be used
as an alternative to the dlaf:: transformMPI mechanism
that uses yield_while to wait on an MPI request.

The completion modes may be set via the environment variable
  PIKA_MPI_COMPLETION_MODE=<numeric value>
which by default will select the one chosen by pika/dlaf
developers known to give good results across a broad range of
use cases.

polling:
The pika polling loop may test for one or multiple request
completions on each iteration through the scheduling loop
the environment var
  PIKA_MPI_MSG_THROTTLE=<numeric value> (default 32)

streams and throttling:
Each MPI request may be launched on a named stream that may
be throttled to reduce the number of messages in flight.
Currently this does not seem to impact performance, but is
supported for future use. The environment variable
  PIKA_MPI_MSG_THROTTLE=<numeric value> (default (off) 65536)
may be used to set the throttling on all streams.
See pika docs for info on how to throttle streams individually.

Use pika::mpi::pool_name instead of raw "mpi" and cleanup some
of the code around the setup of mpi.