Fix waitLocalTiles
Use future::get instead of wait to ensure that the tile contained in the
future is released on the calling thread. When calling wait the result
of the future is ready, but subsequent reads or readwrites of the same
tile are not guaranteed to be immediately ready because the tile itself
can be released after wait returns on another thread.