Fix waitLocalTiles to guarantee that the next access won't yield (#838)
This changes waitLocalTiles to use future::get instead of wait to ensure that the tile contained in the future is released on the calling thread. When calling wait we ensure that the result of the future is ready but we do not ensure that the tile has been released. When only calling wait it's possible for wait to return but the contained tile not to have been released yet. This means that subsequent reads may end up yielding if the tile has not yet been released. In most cases the tile is released very soon after signaling wait but there is a delay which opens up the door to race conditions.