Allow incomplete first tile/block in `Distribution` (#901)
This adds support for having incomplete first tiles in Distribution. This is introduced through a global offset in terms of number of elements. The size of the distribution is relative to the offset.

The offset is allowed to be arbitrarily large (i.e. larger than the block size). If the offset is larger than the block size I normalize it to be smaller than the block size and instead shift the source rank to account for the larger-than-block-size offset.

At the moment I've simply added an optional last offset parameter to all the Distribution constructors to not change all the callsites drastically.